chore: introduce env-utils package
