+++ packages/env-utils/package.json
@@ -1,0 +1,15 @@
+{
+    "name": "@trezor/env-utils",
+    "version": "1.0.0",
+    "private": true,
+    "license": "See LICENSE.md in repo root",
+    "sideEffects": false,
+    "main": "src/index",
+    "scripts": {
+        "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
+        "type-check": "tsc --build"
+    },
+    "dependencies": {
+        "ua-parser-js": "^1.0.2"
+    }
+}
