+++ packages/suite/src/reducers/suite/suiteReducer.ts
@@ -6,7 +6,8 @@ import { SUITE, STORAGE } from '@suite-actions/constants';
 import { DISCOVERY } from '@wallet-actions/constants';
 import { Action, TrezorDevice, Lock, TorStatus } from '@suite-types';
 import type { Locale } from '@suite-config/languages';
-import { isWeb, getWindowWidth } from '@suite-utils/env';
+import { isWeb } from '@suite-utils/env';
+import { getWindowWidth } from '@trezor/env-utils';
 import { ensureLocale } from '@suite-utils/l10n';
 import { getNumberFromPixelString } from '@trezor/utils';
 import type { OAuthServerEnvironment } from '@suite-types/metadata';
