+++ packages/suite/src/utils/suite/env.ts
@@ -1,40 +1,9 @@
-import UAParser from 'ua-parser-js';
 import { desktopApi, SuiteThemeVariant } from '@trezor/suite-desktop-api';
-
+import { getPlatform, isAndroid, isChromeOs } from '@trezor/env-utils';
 import type { EnvironmentType } from '@suite-types';
-
-/* This way, we can override simple utils, which helps to polyfill methods which are not available in react-native. */
-export const getUserAgent = () => window.navigator.userAgent;
-
-// List of platforms https://docker.apachezone.com/blog/74
-export const getPlatform = () => window.navigator.platform;
-
-export const getPlatformLanguages = () => window.navigator.languages;
-
-export const getScreenWidth = () => window.screen.width;
-
-export const getScreenHeight = () => window.screen.height;
-
-export const getWindowWidth = () => window.innerWidth;
-
-export const getWindowHeight = () => window.innerHeight;
-
-export const getLocationOrigin = () => window.location.origin;
-
-export const getLocationHostname = () => window.location.hostname;
 
 /* For usage in Electron */
 export const getProcessPlatform = () => process.platform;
-
-let userAgentParser: UAParser;
-const getUserAgentParser = () => {
-    if (!userAgentParser) {
-        const ua = getUserAgent();
-        userAgentParser = new UAParser(ua);
-    }
-    return userAgentParser;
-};
-
 export const isMacOs = () => {
     if (getProcessPlatform() === 'darwin') return true;
     if (typeof window === 'undefined') return;
@@ -50,10 +19,6 @@ export const isWindows = () => {
 };
 
 export const isIOs = () => ['iPhone', 'iPad', 'iPod'].includes(getPlatform());
-
-export const isAndroid = () => /Android/.test(getUserAgent());
-
-export const isChromeOs = () => /CrOS/.test(getUserAgent());
 
 export const isLinux = () => {
     if (getProcessPlatform() === 'linux') return true;
@@ -75,17 +40,6 @@ export const getOsName = () => {
 
     return '';
 };
-
-/* Not correct for Linux as there is many different distributions in different versions */
-export const getOsVersion = () => getUserAgentParser().getOS().version || '';
-
-export const getBrowserName = (): string => {
-    const browserName = getUserAgentParser().getBrowser().name;
-
-    return browserName?.toLowerCase() || '';
-};
-
-export const getBrowserVersion = (): string => getUserAgentParser().getBrowser().version || '';
 
 export const isWeb = () => process.env.SUITE_TYPE === 'web';
