+++ packages/suite/src/utils/wallet/coinmarket/buyUtils.ts
@@ -5,7 +5,8 @@ import { AppState } from '@suite-types';
 import { AmountLimits } from '@wallet-types/coinmarketBuyForm';
 import { BuyTrade, BuyTradeQuoteRequest, BuyTradeStatus } from 'invity-api';
 import { invityApiSymbolToSymbol } from '@wallet-utils/coinmarket/coinmarketUtils';
-import { getLocationOrigin, isDesktop } from '@suite-utils/env';
+import { isDesktop } from '@suite-utils/env';
+import { getLocationOrigin } from '@trezor/env-utils';
 
 // loop through quotes and if all quotes are either with error below minimum or over maximum, return the limits
 export function getAmountLimits(
