+++ packages/suite/src/views/settings/general/Language.tsx
@@ -10,7 +10,7 @@ import * as languageActions from '@settings-actions/languageActions';
 import { ActionColumn, ActionSelect, SectionItem, TextColumn } from '@suite-components/Settings';
 import { useAnchor } from '@suite-hooks/useAnchor';
 import { SettingsAnchor } from '@suite-constants/anchors';
-import { getPlatformLanguages } from '@suite-utils/env';
+import { getPlatformLanguages } from '@trezor/env-utils';
 
 const onlyComplete = (locale: [string, LocaleInfo]): locale is [Locale, LocaleInfo] =>
     !!locale[1].complete;
